refactor: move data messages to data mod
