+++ src/cache/messages/data/list/list_pop_front.rs
