+++ src/cache/messages/data/list/list_remove_value.rs
