+++ src/cache/messages/data/scalar/delete.rs
