+++ src/cache/messages/data/scalar/item_get_type.rs
