+++ src/cache/messages/data/scalar/key_exists.rs
